Finish SQL methods. Start Spark
