+++ resources/templates/home.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Spark</title>
+</head>
+<body>
+    <p>sup, {{Hello}}</p>
+    <form action="/login" method="post">
+        <input type="text" placeholder="name: " name="name">
+        <button type="submit">login</button>
+    </form>
+</body>
+</html>
